refactor(hocon_schema): allow plain map (not richmap) as for check
This commit adds APIs check_plain/2, check_plain/3 to hocon_schema
module which allows checking plain-map (instead of richmap) as
config input.

The old implementation has a mixed global and partial mutation
of the input config (richmap), so the majority of this commit
is mainly to refactor the map API implementation, in order to
gain a more clear picture of both the shcema stack and the data
stack.