refactor(hocon_schema): avoid deep map merge
The type check implementation is a depth-fist recursion,
i.e. values are checked (and updated) from tree leaves up to to root.

Prior to this commit, when updating a field value, it uses
deep map merge for each tree root, however the sub trees are
already merged. So it's end up doing merges over and over again,
which is logcially OK, but a wast of resource.