fix: better type check exception
prior to this change, all expected fields are hinted in the exception
when there is an unknown field name.
e.g.
input = [a, b, x], schema = [a, b, c]
hint: unknnwon = [x], expected = [a, b, c]

with this fix, the fields which matches the schema are excluded from
the hint
e.g.
input = [a, b, x], schema = [a, b, c]
hint: unknown = [x], expected = [c]