refactor: get rid of the of FROM_ENV_VAR value wrapper
The value wrapper was added to allow palin (non-rich) maps
checks to be albe to carry around this piece of metadata
so it can log the overrides after all checks are done.

This has caused quite some trobules for the value
converter and validator callbacks as they had to
call hocon_tconf:remove_env_meta/1 to remove the metadata.

Going forward, we'll have to accept the fact that if plain
value is used, no override logging will be avaialble.