refactor: turn the hidden (boolean) flag into importance (enum)
prior to this change, the 'hidden' boolean flag was is used to
hide fields in the schema dumps.
now it is changed to use a enum flag named 'importance'

importance can be 'high', 'medium', 'low' or 'hidden'.

if a filed is at 'hidden' importance level, the default values
will not be filled in hocon_tconf:make_serializable.