fix(hocon_token): fix scan error message format
prior to this change, the scan error is formated as a
utf8 binary string, this caused some unreadable error messages
if the error is not formated to console using ~ts.

this commit changes the error report to a map in which
the line number and file paht/name is readable.
this is usually enough to find the error.