fix(schema): Allow unresolvable struct desc
In hocon schema, field desc is allowed to be missing and unresolvable.
Prior to this fix, the struct desc is allowed to be missing but
not unresolvable.
For example, if the desc is resoved from {desc, reference}, but
failed. i.e. returned 'undefined', this atom is used in the iodata
output which then causes 'badarg' error when getting serialized